moved boundary conditions into constants list
it doesn't make any sense to calculate the states
that we know for sure will stay constant. However,
there's still the conflict of initial values in the model
and the initial values in the dataset.

The initial values of the states are defined in both
the model and the dataset. While the initial
concentration of the boundary species are defined
only in the model. so attention is needed to make
sure that they are in consistency, i.e., they are all
from the same steady states.

Further development is need to tackle this problem.
option can be  also define the initial concentration
of the boundary species in the dataset `attr`.

We will see.